requirements.txt
# 基础依赖
fastapi==0.100.0
uvicorn==0.23.0
pydantic==2.0.2
python-dotenv==1.0.0

# API客户端
httpx==0.24.1
requests==2.31.0

# AI/ML相关
openai==1.0.0
numpy==1.24.3
scikit-learn==1.3.0
faiss-cpu==1.7.4  # 如果使用GPU，可替换为faiss-gpu

# 数据存储与检索
elasticsearch==8.9.0
pandas==2.0.3

# 向量处理
sentence-transformers==2.2.2  # 用于生成文本嵌入向量

# 用户服务
pyjwt==2.8.0
email-validator==2.0.0

# 课标知识体系
networkx==3.1  # 用于知识图谱结构分析
pygraphviz==1.10  # 用于知识图谱可视化 (可选，需要系统安装graphviz)

# 测试工具
pytest==7.4.0
pytest-cov==4.1.0
unittest-xml-reporting==3.2.0  # XML格式测试报告生成
html-testRunner==1.2.1  # HTML格式测试报告生成

# 集成测试工具
# TestClient由fastapi提供，不需要单独安装
coverage==7.2.7  # 代码覆盖率分析
faker==19.3.0  # 生成测试数据

# 开发工具
black==23.7.0
flake8==6.1.0
isort==5.12.0 